Core: Updated StudentTest with StudentGroupId and updated TestStatisticServises method with optional parameter groupId

+++ TestingSystem/Applications/WebClient/Requests/CreateTestStatisticRequest.cs
@@ -12,7 +12,7 @@ namespace Applications.WebClient.Requests
         [DataMember]
         public short TestId { get; set; }
         [DataMember]
-        public int? GroupId { get; set; }
+        public short? GroupId { get; set; }
         public DateTime? CreationDate { get; set; }
     }
 }
